style: satisfy the perfectionist dylint lints in the stage code
